Lab-3: Add exception handling

+++ WebServices/src/com/dolganova/lab1/BeautyProductWebService.java
@@ -4,6 +4,11 @@ package com.dolganova.lab1;
  * Created by Polina on 23.02.19.
  */
 
+import com.dolganova.exception.EmptyIdentifierException;
+import com.dolganova.exception.EmptyIdentifierExceptionBean;
+import com.dolganova.exception.NonKeySpecifiedException;
+import com.dolganova.exception.NonKeySpecifiedExceptionBean;
+
 import java.util.List;
 import javax.jws.WebMethod;
 import javax.jws.WebService;
@@ -19,28 +24,65 @@ public class BeautyProductWebService {
     }
 
     @WebMethod(operationName = "insertBeautyProduct")
-    public Integer insertBeautyProduct(String name, String producingСountry, String vendorСode, String category, String price) {
+    public Integer insertBeautyProduct(String name, String producingСountry, String vendorСode, String category, String price) throws NonKeySpecifiedException {
+        if ((name == null || name.isEmpty()) &&
+            (producingСountry == null || producingСountry.isEmpty()) &&
+            (vendorСode == null || vendorСode.isEmpty()) &&
+            (category == null || category.isEmpty())&&
+            (price == null || price.isEmpty()))
+        {
+            NonKeySpecifiedExceptionBean fault = new NonKeySpecifiedExceptionBean();
+            throw new NonKeySpecifiedException("No one key is specified.", fault);
+        }
         PostgreSQLDAO dao = new PostgreSQLDAO();
         Integer beautyProductId = dao.insertBeautyProduct(name, producingСountry, vendorСode, category, price);
         return beautyProductId;
     }
 
     @WebMethod(operationName = "updateBeautyProduct")
-    public String updateBeautyProduct(String id, String name, String producingСountry, String vendorСode, String category, String price) {
+    public String updateBeautyProduct(String id, String name, String producingСountry, String vendorСode, String category, String price) throws EmptyIdentifierException, NonKeySpecifiedException {
+        if (id == null || id.isEmpty())
+        {
+            EmptyIdentifierExceptionBean fault = new EmptyIdentifierExceptionBean();
+            throw new EmptyIdentifierException("Product identifier is not specified.", fault);
+        }
+        if ((name == null || name.isEmpty()) &&
+                (producingСountry == null || producingСountry.isEmpty()) &&
+                (vendorСode == null || vendorСode.isEmpty()) &&
+                (category == null || category.isEmpty())&&
+                (price == null || price.isEmpty()))
+        {
+            NonKeySpecifiedExceptionBean fault = new NonKeySpecifiedExceptionBean();
+            throw new NonKeySpecifiedException("No one key is specified.", fault);
+        }
         PostgreSQLDAO dao = new PostgreSQLDAO();
         String status = dao.updateBeautyProduct(id, name, producingСountry, vendorСode, category, price) ;
         return status;
     }
 
     @WebMethod(operationName = "deleteBeautyProduct")
-    public String deleteBeautyProduct(String id) {
+    public String deleteBeautyProduct(String id) throws EmptyIdentifierException {
+        if (id == null || id.isEmpty())
+        {
+            EmptyIdentifierExceptionBean fault = new EmptyIdentifierExceptionBean();
+            throw new EmptyIdentifierException("Product identifier is not specified.", fault);
+        }
         PostgreSQLDAO dao = new PostgreSQLDAO();
         String status = dao.deleteBeautyProduct(id) ;
         return status;
     }
 
     @WebMethod(operationName = "findBeautyProduct")
-    public List<BeautyProduct> findBeautyProduct(String id, String name, String producingСountry, String vendorСode, String category, String price) {
+    public List<BeautyProduct> findBeautyProduct(String id, String name, String producingСountry, String vendorСode, String category, String price) throws NonKeySpecifiedException {
+        if ((name == null || name.isEmpty()) &&
+                (producingСountry == null || producingСountry.isEmpty()) &&
+                (vendorСode == null || vendorСode.isEmpty()) &&
+                (category == null || category.isEmpty())&&
+                (price == null || price.isEmpty()))
+        {
+            NonKeySpecifiedExceptionBean fault = new NonKeySpecifiedExceptionBean();
+            throw new NonKeySpecifiedException("No one key is specified.", fault);
+        }
         PostgreSQLDAO dao = new PostgreSQLDAO();
         List<BeautyProduct> beautyProducts = dao.findBeautyProduct(id, name, producingСountry, vendorСode, category, price) ;
         return beautyProducts;
